requirements.txt
pandas
numpy
scikit-learn
matplotlib
streamlit
Pillow
google.cloud
s3fs
gcsfs
